feat(gha): publish developer package
Introduces a new GitHub Actions workflow `publish-dev-package.yaml`
which publishes a developer package to the GitHub npm registry
whenever a commit is pushed to the `main` branch. A developer package
has a special version string, version number followed by a dash (`-`)
plus the short commit hash of the commit used to build the package).

`publish-dev-package.yaml` calls the `publish-package.yaml` workflow to
do the actual job.